+++ contracts/samples/LegacyTokenPaymaster.sol
@@ -32,7 +32,7 @@ contract LegacyTokenPaymaster is BasePaymaster, ERC20 {
         _mint(address(this), 1);
 
         //owner is allowed to withdraw tokens from the paymaster's balance
-        _approve(address(this), msg.sender, type(uint).max);
+        _approve(address(this), msg.sender, type(uint256).max);
     }
 
 
@@ -55,7 +55,7 @@ contract LegacyTokenPaymaster is BasePaymaster, ERC20 {
         _approve(address(this), owner(), 0);
         super.transferOwnership(newOwner);
         // new owner is allowed to withdraw tokens from the paymaster's balance
-        _approve(address(this), newOwner, type(uint).max);
+        _approve(address(this), newOwner, type(uint256).max);
     }
 
     //Note: this method assumes a fixed ratio of token-to-eth. subclass should override to supply oracle
@@ -102,7 +102,7 @@ contract LegacyTokenPaymaster is BasePaymaster, ERC20 {
      * the user's TX , back to the state it was before the transaction started (before the validatePaymasterUserOp),
      * and the transaction should succeed there.
      */
-    function _postOp(PostOpMode mode, bytes calldata context, uint256 actualGasCost, uint actualUserOpFeePerGas) internal override {
+    function _postOp(PostOpMode mode, bytes calldata context, uint256 actualGasCost, uint256 actualUserOpFeePerGas) internal override {
         //we don't really care about the mode, we just pay the gas with the user's tokens.
         (mode);
         address sender = abi.decode(context, (address));
